# api/openapi.json
{
  "components": {
    "schemas": {
...
      "ClaimCreate": {
        "properties": {
          "claim_type": {
            "$ref": "#/components/schemas/ClaimType"
          },
          "place_id": {
            "format": "uuid",
            "title": "Place Id",
            "type": "string"
          },
          "scope": {
            "$ref": "#/components/schemas/ClaimScope",
            "default": "ALL_MENU"
          }
        },
        "required": [
          "place_id",
          "claim_type"
        ],
        "title": "ClaimCreate",
        "type": "object"
      },
...
      "ClaimRead": {
        "properties": {
          "claim_type": {
            "$ref": "#/components/schemas/ClaimType"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "place_id": {
            "format": "uuid",
            "title": "Place Id",
            "type": "string"
          },
          "scope": {
            "$ref": "#/components/schemas/ClaimScope"
          },
          "status": {
            "$ref": "#/components/schemas/ClaimStatus"
          }
        },
        "required": [
          "id",
          "place_id",
          "claim_type",
          "scope",
          "status",
          "expires_at",
          "created_at"
        ],
        "title": "ClaimRead",
        "type": "object"
      },
      "ClaimScope": {
        "enum": [
          "ALL_MENU",
          "SPECIFIC_ITEMS"
        ],
        "title": "ClaimScope",
        "type": "string"
      },
...
      "ClaimType": {
        "enum": [
          "ZABIHA",
          "HALAL_CHICKEN_ONLY",
          "PORK_FREE",
          "NO_ALCOHOL",
          "HALAL_MEAT_AVAILABLE"
        ],
        "title": "ClaimType",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "InviteInfoResponse": {
        "description": "Prefetch response for GET /auth/invite/{token}.\n\nThe set-password page calls this to render \"Set your password for\n<email>\" before the user submits. We expose the minimum identifying\ninfo (email + display_name) and nothing else \u2014 no role, no user id,\nno created-at \u2014 so the endpoint can't be used as an oracle for\nenumerating the user table.",
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "title": "InviteInfoResponse",
        "type": "object"
      },
      "LoginRequest": {
        "additionalProperties": false,
        "description": "POST /auth/login body.",
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
      "LoginResponse": {
        "description": "POST /auth/login response.\n\n``redirect_path`` is computed server-side per user role so the admin\npanel, future owner dashboard, etc. land the user in the right\nplace without hard-coding role-aware routing on every client.",
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "redirect_path": {
            "title": "Redirect Path",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "email",
          "role",
          "redirect_path"
        ],
        "title": "LoginResponse",
        "type": "object"
      },
...
      "PlaceCreate": {
        "properties": {
          "address": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "lat": {
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Lat",
            "type": "number"
          },
          "lng": {
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Lng",
            "type": "number"
          },
          "name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "lat",
          "lng"
        ],
        "title": "PlaceCreate",
        "type": "object"
      },
...
      "PlaceSearchResult": {
        "properties": {
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "country_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country Code"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "lat": {
            "title": "Lat",
            "type": "number"
          },
          "lng": {
            "title": "Lng",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          }
        },
        "required": [
          "id",
          "name",
          "address",
          "lat",
          "lng"
        ],
        "title": "PlaceSearchResult",
        "type": "object"
      },
...
      "SetPasswordRequest": {
        "additionalProperties": false,
        "description": "POST /auth/set-password body.\n\n``password`` minimum is 8 characters to match the login flow's\ncurrent expectation. No max-complexity rules right now \u2014 we can\nlayer a zxcvbn check later if the risk profile warrants it.",
        "properties": {
          "password": {
            "maxLength": 256,
            "minLength": 8,
            "title": "Password",
            "type": "string"
          },
          "token": {
            "maxLength": 128,
            "minLength": 16,
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token",
          "password"
        ],
        "title": "SetPasswordRequest",
        "type": "object"
      },
      "SetPasswordResponse": {
        "description": "POST /auth/set-password response.\n\nReturns the same shape as LoginResponse \u2014 setting a password via\ninvite auto-logs the user in, so the client treats the two\nresponses identically: store whatever ``/me`` cache is needed and\nroute to ``redirect_path``.",
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "redirect_path": {
            "title": "Redirect Path",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "email",
          "role",
          "redirect_path"
        ],
        "title": "SetPasswordResponse",
        "type": "object"
      },
...
      "UserRole": {
        "enum": [
          "ADMIN",
          "OWNER",
          "VERIFIER",
          "CONSUMER"
        ],
        "title": "UserRole",
        "type": "string"
      },
...
    }
  },
  "info": {
    "title": "TrustHalal API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/auth/invite/{token}": {
      "get": {
        "description": "Prefetch an invite so the set-password page can show context.\n\nReturns the invited user's email + display_name so the UI can\nrender \"Set your password for rozebm@example.com\" before the\nuser submits. Does NOT consume the token \u2014 that happens in\n``POST /auth/set-password``.\n\nFailure modes all land on the same generic 400 to avoid becoming\nan oracle: invalid, expired, and already-consumed tokens are\nindistinguishable from the outside. The client just sees\nINVITE_INVALID and tells the user the link needs replacing.",
        "operationId": "get_invite_info_auth_invite__token__get",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteInfoResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Invite Info",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "description": "Email + password login. Sets the session cookie on success.\n\nSecurity notes:\n  * Single generic error (\"Invalid email or password\") on any\n    failure path \u2014 no user enumeration. The ``UnauthorizedError``\n    wrapper keeps the error code stable for client-side branching\n    without leaking which of email/password/active/has-password\n    failed.\n  * Password verification runs even when the user isn't found, to\n    keep timing roughly similar. We do this with a dummy hash that\n    matches the cost of a real one.",
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "description": "Invalidate the current session + clear the cookie.\n\nIdempotent: no cookie / malformed cookie / already-revoked session\nall 204 cleanly. The client is losing the cookie either way, so\nsurfacing errors would only confuse \"sign out\" UX.",
        "operationId": "logout_auth_logout_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "tht_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hht Session"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Logout",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/set-password": {
      "post": {
        "description": "Burn an invite token + set the user's password + sign them in.\n\nSingle transaction:\n  1. Resolve the token (same generic failure surface as\n     ``get_invite_info``).\n  2. Hash the new password with argon2id and set it on the user.\n  3. Mark the token consumed.\n  4. Revoke every other outstanding session for this user (covers\n     the \"re-invite after lost password\" case: any live sessions\n     for the old password are kicked out \u2014 the user only wants\n     the freshly-made one).\n  5. Create a new session + set the cookie.\n\nAnyone who has a valid invite can set the password, by design \u2014\nthat's the whole point. We don't require re-auth with the old\npassword because in the invite flow there may be no old password\nat all.",
        "operationId": "set_password_with_invite_auth_set_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetPasswordResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Password With Invite",
        "tags": [
          "auth"
        ]
      }
    },
    "/claims": {
      "post": {
        "operationId": "post_claim_claims_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "in": "cookie",
            "name": "tht_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hht Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Claim",
        "tags": [
          "claims"
        ]
      }
    },
...
    "/me": {
      "get": {
        "operationId": "get_me_me_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "in": "cookie",
            "name": "tht_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hht Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Me",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/places": {
      "get": {
        "operationId": "search_places_places_get",
        "parameters": [
          {
            "in": "query",
            "name": "lat",
            "required": true,
            "schema": {
              "maximum": 90,
              "minimum": -90,
              "title": "Lat",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "lng",
            "required": true,
            "schema": {
              "maximum": 180,
              "minimum": -180,
              "title": "Lng",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "radius",
            "required": true,
            "schema": {
              "exclusiveMinimum": 0,
              "maximum": 100000,
              "title": "Radius",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "exclusiveMinimum": 0,
              "maximum": 200,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlaceSearchResult"
                  },
                  "title": "Response Search Places Places Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Places",
        "tags": [
          "Places"
        ]
      },
      "post": {
        "operationId": "post_place_places_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "in": "cookie",
            "name": "tht_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hht Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaceCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Place",
        "tags": [
          "Places"
        ]
      }
    },
...
  }
}